backend/requirements.txt
django==4.2.11
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-phonenumber-field[phonenumbers]==7.3.0

# testing
pytest==8.1.1
pytest-django==4.8.0
